refactor: organize accounts api methods
