+++ src/api/accounts/cash/delete_cash_account.py
