+++ walter.py
@@ -1,6 +1,6 @@
-from src.api.credit_accounts.create_credit_account import CreateCreditAccount
-from src.api.credit_accounts.delete_credit_account import DeleteCreditAccount
-from src.api.credit_accounts.get_credit_accounts import GetCreditAccounts
+from src.api.accounts.credit_accounts import CreateCreditAccount
+from src.api.accounts.credit_accounts import DeleteCreditAccount
+from src.api.accounts.credit_accounts import GetCreditAccounts
 from src.api.plaid.create_link_token import CreateLinkToken
 from src.api.plaid.exchange_public_token import ExchangePublicToken
 from src.api.plaid.refresh_transactions import RefreshTransactions
@@ -9,13 +9,13 @@ from src.api.transactions.add_transaction import AddTransaction
 from src.api.add_stock import AddStock
 from src.api.auth_user import AuthUser
 from src.api.change_password import ChangePassword
-from src.api.create_cash_account import CreateCashAccount
+from src.api.accounts.cash.create_cash_account import CreateCashAccount
 from src.api.create_user import CreateUser
-from src.api.delete_cash_account import DeleteCashAccount
+from src.api.accounts.cash.delete_cash_account import DeleteCashAccount
 from src.api.delete_stock import DeleteStock
 from src.api.transactions.delete_transaction import DeleteTransaction
 from src.api.transactions.edit_transaction import EditTransaction
-from src.api.get_cash_accounts import GetCashAccounts
+from src.api.accounts.cash.get_cash_accounts import GetCashAccounts
 from src.api.get_news_summary import GetNewsSummary
 from src.api.get_newsletter import GetNewsletter
 from src.api.get_newsletters import GetNewsletters
@@ -32,7 +32,7 @@ from src.api.send_newsletter import SendNewsletter
 from src.api.send_verify_email import SendVerifyEmail
 from src.api.subscribe import Subscribe
 from src.api.unsubscribe import Unsubscribe
-from src.api.update_cash_account import UpdateCashAccount
+from src.api.accounts.cash.update_cash_account import UpdateCashAccount
 from src.api.update_password import UpdatePassword
 from src.api.update_user import UpdateUser
 from src.api.verify_email import VerifyEmail
